Moved CassandraClient to .cassandra subpackage.
