+++ src/main/scala/ztis/UserAndRatingRepository.scala
@@ -1,6 +1,7 @@
 package ztis
 
 import com.datastax.driver.core.Row
+import ztis.cassandra.CassandraClient
 
 import scala.collection.JavaConverters._
